wip: modify recommendation logic
